// frontend/src/lib/api.d.ts
export interface components {
    schemas: {
// ...
        StreamCreate: {
            description?: string | null;
            discard?: components["schemas"]["Discard"];
            /** @default 120 */
            duplicate_window_seconds?: number;
            /** @default 0 */
            max_age_seconds?: number;
            /** @default -1 */
            max_bytes?: number;
            /** @default -1 */
            max_msg_size?: number;
            /** @default -1 */
            max_msgs?: number;
            name: string;
            /** @default 1 */
            replicas?: number;
            retention?: components["schemas"]["Retention"];
            storage?: components["schemas"]["Storage"];
            subjects: string[];
        };
        /** StreamDetail */
        StreamDetail: {
            cluster: components["schemas"]["ClusterInfo"];
            /** Format: date-time */
            created: string;
            description: string | null;
            limits: components["schemas"]["StreamLimits"];
            mirror?: string | null;
            name: string;
            /** @default [] */
            placement_tags?: string[];
            replicas: number;
            republish_to?: string | null;
            retention: components["schemas"]["Retention"];
            /** @default false */
            sealed?: boolean;
            /** @default [] */
            sources?: string[];
            state: components["schemas"]["StreamState"];
            storage: components["schemas"]["Storage"];
            subjects: string[];
        };
// ...
        StreamUpdate: {
            allow_rollup?: boolean | null;
            deny_delete?: boolean | null;
            deny_purge?: boolean | null;
            description?: string | null;
            discard?: components["schemas"]["Discard"] | null;
            duplicate_window_seconds?: number | null;
            max_age_seconds?: number | null;
            max_bytes?: number | null;
            max_consumers?: number | null;
            max_msg_size?: number | null;
            max_msgs?: number | null;
            max_msgs_per_subject?: number | null;
            metadata?: {
                [key: string]: string;
            } | null;
            replicas?: number | null;
            republish_destination?: string | null;
            republish_source?: string | null;
            subjects?: string[] | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}